+++ src/main/java/ladder/domain/LadderStrategy.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface LadderStrategy {
     Direction direction();
