+++ src/main/java/ladder/domain/Players.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.List;
 import java.util.Objects;
